requirements.txt
rake-nltk 
numpy 
googlesearch-python
